update playground and rename map folder
