+++ src/components/Maps/index.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import Map from './Map';
+
+export default ({ config }) => {
+  return (
+    <>
+      {config.map(map => {
+        return <Map config={map} />;
+      })}
+    </>
+  );
+};
