+++ src/pages/playground/karte.js
@@ -6,7 +6,7 @@ import {
   Section,
   SectionInner,
 } from '../../components/Layout/Sections';
-import Map from '../../components/Map';
+import Map from '../../components/Maps/Map';
 
 export default () => {
   return (
@@ -17,7 +17,14 @@ export default () => {
       <SectionWrapper>
         <Section title="karte">
           <SectionInner wide={true}>
-            <Map state="schleswig-holstein" />
+            <Map
+              config={{
+                state: 'berlin',
+                config: {
+                  bounds: [1, 2, 3, 4],
+                },
+              }}
+            />
           </SectionInner>
         </Section>
       </SectionWrapper>
